Don't save MainActivity's view state.
The views are set up to match the current state of the AudioBookManager
and the PlaybackService.
This change should also fix IllegalStateExceptions thrown by the
FragmentManager's state being changed after its state is saved
(this issue could be fixed by making sure thet showPage() is only called
at specific points in the Activity's lifecycle but now that view state
is no longer saved there's no need to fix this).