Fix EventBus-related crash on start on API pre-21.
Overriding a method with ParcelableBundle argument
(i.e. onSaveInstanceStace) causes reflection code inside EventBus to
throw an execption on Android API20 and lower. The reason is that the
ParcelableBundle class has been introduec in API21 and is not available
in earlier versions.

The workaround for this issue is to register an object of an inner class
that doesn't have arguments of unavailable classes.