Don't store state in enums.
This is a weird guess fix.
Crash-log data suggests that sometimes there are two active instances of
the MainActivity (active meaning between onStart and onStop). This could
be some kind of split-screen or multi-app feature (most but not all
crashes come from either Samsung tablets or rooted devices which would
support this theory).
Most of the UI shouldn't care how many instances are running. One
notable exception was the UiControllerMain.State enum whose (global)
instances had local state that would change in time. With this commit
the UiControllerMain.State becomes a regular class so there should be no
problems with multiple instances of the UiControllerMain object.
Whether this indicates a bigger bug that needs fixing or just a feature
of some Android versions that needs to be accepted remains to be seen.