Don't leak the AudioTracks from SoundBank.
An AudioTrack needs to be released, otherwise it use resources and
after a limit is reached it is not possible to create any new tracks
thus making it impossible to play music.
The SoundBank object was being created by Dagger each time playback
started (each time ClassicPlaybackUi was created). Now SoundBank is
a singleton so it is only created once during the lifetime of the
process.