Fix a race condition crash when all books are removed.
When the screen is turned on the UiControllerBookList.screenOnReceiver
is triggered to read the current book title. The app also updates the
internal book list when the Activity is started. If there are no books
the UiControllerBookList.screenOnReceiver is unregistered.
However, it is possible that the system has already posted a call to
its onReceive() method that then crashes because there is no current
book to get the title of.