Set initial UI state in onResumeFragments.
onResumeFragments is when the fragments state has been restored.
The app doesn't save any UI state (onSaveInstanceState implementations
are empty, and all UI state is derived from the state of the services
for playback and sample downloads) it's better this way in case some
operations are performed at this stage.