Don't crash after playback is stopped while rewinding.
Stopping rewind accesses the currently played audiobook in order to set
the position. This throws an exception because playback is already
stopped. The exception is being caught by EventBus but leaves the player
in an inconsistent state leading to a crash later on.

This fix aborts rewind gracefully in this case.

One way to trigger this behaviour is to keep the device relatively
steady with the screen facing down while rewinding. This can be a valid
use-case for a person laying in bed and the correct fix is probably to
disable face-down detection when rewinding. Doing this requires some
refactoring and is planned together with other improvements.